requirements.txt
asgiref==3.7.2
Brotli==1.1.0
click==8.1.7
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
h11==0.14.0
inflection==0.5.1
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
pystache==0.6.5
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
static==1.1.1
tzdata==2023.4
uritemplate==4.1.1
uvicorn==0.27.1
whitenoise==6.6.0
